+++ Tests/TaskHelp.test.ps1
@@ -39,7 +39,7 @@ task Test {
 	Invoke-Build -File z.ps1 -WhatIf
 	equals $log.Value "Task 'Test': unknown parameter 'Missing.'."
 
-	$r = Show-TaskHelp . z.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 . z.ps1 -Format {$args[0]}
 
 	equals $r.Task.Count 1
 	equals $r.Task[0].Name Test
@@ -71,7 +71,7 @@ task UndocumentedStuff {
 	Invoke-Build -File ../.build.ps1 -WhatIf
 
 	# test default task call with code
-	$r = Show-TaskHelp '' ../.build.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 '' ../.build.ps1 -Format {$args[0]}
 	equals $r.Task.Count 1
 	equals $r.Task[0].Name .
 	equals $r.Task[0].Synopsis 'The default task: make, test, clean.'
@@ -88,7 +88,7 @@ task UndocumentedStuff {
 	equals $r.Environment.Count 0
 
 	# test a task with code and environment variable
-	$r = Show-TaskHelp test ../.build.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 test ../.build.ps1 -Format {$args[0]}
 	equals $r.Task.Count 1
 	equals $r.Task[0].Name test
 	equals $r.Task[0].Synopsis 'Test and check expected output.'
@@ -106,7 +106,7 @@ task UndocumentedStuff {
 	equals $r.Environment[0] MERGE
 
 	# call with -NoCode
-	$r = Show-TaskHelp test ../.build.ps1 -Format {$args[0]} -NoCode
+	$r = Show-TaskHelp.ps1 test ../.build.ps1 -Format {$args[0]} -NoCode
 	equals $r.Parameters.Count 0
 	equals $r.Environment.Count 0
 }
@@ -126,7 +126,7 @@ task SkipAssignment {
 	Invoke-Build . z.ps1 -WhatIf
 
 	# call and test
-	$r = Show-TaskHelp . z.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 . z.ps1 -Format {$args[0]}
 	equals $r.Parameters.Count 0
 	equals $r.Environment.Count 0
 
@@ -145,7 +145,7 @@ task TwoTasks {
 	Invoke-Build build, release z.ps1 -WhatIf
 
 	# call and test
-	$r = Show-TaskHelp build, release z.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 build, release z.ps1 -Format {$args[0]}
 	equals $r.Task.Count 2
 	equals $r.Task[0].Name Build
 	equals $r.Task[1].Name Release
@@ -168,7 +168,7 @@ task ForkWithIf {
 	Invoke-Build t1 z.ps1 -WhatIf
 
 	# call and test
-	$r = Show-TaskHelp t1 z.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 t1 z.ps1 -Format {$args[0]}
 	equals $r.Task.Count 1
 	equals $r.Task[0].Name t1
 	equals $r.Jobs.Count 2
@@ -189,7 +189,7 @@ task IfInputsOutputs {
 	Invoke-Build t1 z.ps1 -WhatIf
 
 	# call and test
-	$r = Show-TaskHelp t1 z.ps1 -Format {$args[0]}
+	$r = Show-TaskHelp.ps1 t1 z.ps1 -Format {$args[0]}
 	equals $r.Parameters.Count 3
 	equals $r.Parameters[0].Name If
 	equals $r.Parameters[1].Name Inputs